backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib
python-jose
python-dotenv
